+++ esperanto-tools-libs/include/runtime/Types.h
@@ -227,6 +227,7 @@ struct DeviceProperties {
   ArchRevision deviceArch_;             ///< device architecture revision
   FormFactor formFactor_;               ///< device form factor
   uint8_t tdp_;                         ///< TDP in Watts
+  uint64_t p2pBitmap_ = 0;              ///< indicates p2p capabilities between this device and others
 };
 
 // NOTE: this is copied directly from device firmware "encoder.h"; we need to find a proper solution. So this will be in