+++ esperanto-tools-libs/src/RuntimeImp.h
@@ -74,6 +74,10 @@ public:
                                const CmaCopyFunction& cmaCopyFunction) final;
   EventId doMemcpyDeviceToHost(StreamId stream, MemcpyList memcpyList, bool barrier,
                                const CmaCopyFunction& cmaCopyFunction) final;
+  EventId doMemcpyDeviceToDevice(StreamId streamSrc, DeviceId deviceDst, const std::byte* d_src, std::byte* d_dst,
+                                 size_t size, bool barrier) final;
+  EventId doMemcpyDeviceToDevice(DeviceId deviceSrc, StreamId streamDst, const std::byte* d_src, std::byte* d_dst,
+                                 size_t size, bool barrier) final;
 
   bool doWaitForEvent(EventId event, std::chrono::seconds timeout = std::chrono::hours(24)) final;
   bool doWaitForStream(StreamId stream, std::chrono::seconds timeout = std::chrono::hours(24)) final;
@@ -89,6 +93,8 @@ public:
   std::vector<StreamError> doRetrieveStreamErrors(StreamId stream) final;
 
   DmaInfo doGetDmaInfo(DeviceId deviceId) const final;
+
+  bool doIsP2PEnabled(DeviceId one, DeviceId other) const final;
 
   ~RuntimeImp() final;